show and search product

+++ database/migrations/2022_05_15_000222_add_visitors_column_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->integer('visitors')->default(0)->after('image');
+            $table->bigInteger('visitors')->default(0)->after('image');
         });
     }
